Update the class Rectangle by adding validation of all setter methods and instantiation (id excluded):
If the input is not an integer, raise the TypeError exception with the
message: <name of the attribute> must be an integer. Example: width must
be an integer
If width or height is under or equals 0, raise the ValueError exception
with the message: <name of the attribute> must be > 0. Example: width
must be > 0
If x or y is under 0, raise the ValueError exception with the message:
<name of the attribute> must be >= 0. Example: x must be >= 0